mcp-servers/ai-os-gateway/requirements.txt
fastmcp>=3.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
asyncpg>=0.30.0
google-cloud-secret-manager>=2.21.0
pydantic>=2.10.0
python-dotenv>=1.0.0

# Phase 2 — Google API clients (install now, use later)
google-api-python-client>=2.160.0
google-auth>=2.38.0
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0
